File size tracking.
The size of the file is stored in the first 4k block on disk.
ReadAt/WriteAt offsets are transparently shifted to operate after the
"metadata" block at sector 0.

File size updates are done when the write to the corresponding block
that extended the file is marked as completed. This causes a dependency
between reads and writes, forcing ReadAt to call Sync before checking
if the read is past the end of the file. If Sync is not called at the
start of write, the write updating the file size could be queued after
the read, extending the window of time a crash could occur and cause
data to be lost.

If a larger window of possible inconsistency is allowed, careful
tracking of the in-memory file size and the file size on disk could
be used to allow file size updates to be queued after read requests.